feat(jepa): trajectory ring buffer + fix boot alignment fault
Add anx_jepa_traj_entry ring buffer (64 entries) in jepa_traj.c that
records (obs, action) pairs during IBAL runs for offline training export.
Dynamic action lists now read from the active world profile rather than
hardcoded names.

Fix anx_memcpy and anx_memset to check pointer alignment before using
8-byte word copies — the unguarded uint64_t* cast caused an alignment
fault at boot when new BSS data shifted .rodata string addresses off
8-byte boundaries (AArch64 SCTLR_EL1.A strict alignment).

Also convert lambda_decay from float to uint16_t Q8 (256=1.0) to
eliminate undefined soft-float ABI builtins on aarch64-none-elf, and
bump version to 2026.4.29.

+++ kernel/include/anx/rlm.h
@@ -64,7 +64,7 @@ struct anx_rlm_config {
 	uint32_t max_steps;		/* hard cap on inference iterations */
 	uint32_t max_tokens;		/* per call (step 0 baseline) */
 	uint32_t min_tokens;		/* lower bound after lambda decay */
-	float lambda_decay;		/* lambda-RLM token decay per infer step */
+	uint16_t lambda_decay;		/* lambda-RLM token decay, Q8: 256=1.0, e.g. 210≈0.82 */
 	bool persist_trace;		/* finalize trace as State Object */
 	bool admit_responses;		/* admit responses to memplane L1 */
 	enum anx_sched_priority priority;
